+++ packages/server/src/subrouteHandlers/purchaseSessionHandlers.ts
@@ -1,23 +1,49 @@
 import { FlowgladActionKey, HTTPMethod } from '@flowglad/shared'
-import { SubRouteHandler } from './types'
+import { SubRouteHandler, SubRouteHandlerResultData } from './types'
 import { FlowgladServer } from '../flowgladServer'
+import { parseErrorStringToErrorObject } from '../serverUtils'
 
 const createPurchaseSession: SubRouteHandler<
   FlowgladActionKey.CreatePurchaseSession
 > = async (params, flowgladServer: FlowgladServer) => {
+  let error:
+    | { code: string; json: Record<string, unknown> }
+    | undefined
+  let status: number
+  let data: SubRouteHandlerResultData<FlowgladActionKey.CreatePurchaseSession> =
+    {}
   if (params.method !== HTTPMethod.POST) {
+    error = {
+      code: 'Method not allowed',
+      json: {},
+    }
+    status = 405
     return {
-      data: {},
-      status: 405,
-      error: 'Method not allowed',
+      data,
+      status,
+      error,
     }
   }
-  const purchaseSession = await flowgladServer.createPurchaseSession(
-    params.data
-  )
+  try {
+    const purchaseSession =
+      await flowgladServer.createPurchaseSession(params.data)
+    data = purchaseSession
+    status = 200
+  } catch (e) {
+    if (e instanceof Error) {
+      error = parseErrorStringToErrorObject(e.message)
+    } else {
+      error = {
+        code: 'Unknown error',
+        json: {},
+      }
+    }
+    status = 500
+  }
   return {
-    data: purchaseSession,
-    status: 200,
+    data,
+    status,
+    error,
   }
 }