+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @flowglad/shared
 
+## 0.4.9
+
+### Patch Changes
+
+- Pass through structured error messages to client
+
+## 0.4.8
+
+### Patch Changes
+
+- Add getRequestingCustomerProfile as fallback for getSessionFromParams
+
 ## 0.4.7
 
 ### Patch Changes